requirements.txt
torch 
torchvision 
torchaudio
torchinfo
torcheval
tensorboard
pandas 
dvc
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
-e .